+++ src/main/kotlin/nexters/linkllet/member/dto/MemberDtos.kt
@@ -1,4 +1,4 @@
-package nexters.linkllet.dto
+package nexters.linkllet.member.dto
 
 data class MemberSignUpRequest(
         val deviceId: String,
