+++ src/main/kotlin/nexters/linkllet/member/presentation/MemberController.kt
@@ -1,7 +1,7 @@
-package nexters.linkllet.controller
+package nexters.linkllet.member.presentation
 
-import nexters.linkllet.dto.MemberSignUpRequest
-import nexters.linkllet.service.MemberService
+import nexters.linkllet.member.dto.MemberSignUpRequest
+import nexters.linkllet.member.service.MemberService
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
@@ -11,12 +11,12 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 @RequestMapping("/api/v1")
 class MemberController(
-        private val memberService: MemberService,
+    private val memberService: MemberService,
 ) {
 
     @PostMapping("/members")
     fun signUp(
-            @RequestBody request: MemberSignUpRequest,
+        @RequestBody request: MemberSignUpRequest,
     ): ResponseEntity<Unit> {
         memberService.signUp(request.deviceId)
         return ResponseEntity.ok().build()
